Introduced AddTargetOverrideFlags() function on base options struct Moved target flags completion logic from root command to base options
