+++ pkg/cmd/cmd.go
@@ -7,7 +7,6 @@ SPDX-License-Identifier: Apache-2.0
 package cmd
 
 import (
-	"context"
 	"flag"
 	"fmt"
 	"os"
@@ -19,7 +18,6 @@ import (
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 	"github.com/spf13/viper"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	cliflag "k8s.io/component-base/cli/flag"
 	"k8s.io/klog/v2"
 
@@ -192,51 +190,6 @@ func initConfig(f *util.FactoryImpl) {
 	cobra.CheckErr(err)
 
 	f.TargetFile = filepath.Join(f.SessionDirectory, targetFilename)
-}
-
-func registerCompletionFuncForGlobalFlags(cmd *cobra.Command, f *util.FactoryImpl, ioStreams util.IOStreams) {
-	utilruntime.Must(cmd.RegisterFlagCompletionFunc("garden", completionWrapper(f, ioStreams, gardenFlagCompletionFunc)))
-	utilruntime.Must(cmd.RegisterFlagCompletionFunc("project", completionWrapper(f, ioStreams, projectFlagCompletionFunc)))
-	utilruntime.Must(cmd.RegisterFlagCompletionFunc("seed", completionWrapper(f, ioStreams, seedFlagCompletionFunc)))
-	utilruntime.Must(cmd.RegisterFlagCompletionFunc("shoot", completionWrapper(f, ioStreams, shootFlagCompletionFunc)))
-}
-
-type cobraCompletionFunc func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective)
-type cobraCompletionFuncWithError func(ctx context.Context, manager target.Manager) ([]string, error)
-
-func completionWrapper(f *util.FactoryImpl, ioStreams util.IOStreams, completer cobraCompletionFuncWithError) cobraCompletionFunc {
-	return func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
-		manager, err := f.Manager()
-
-		if err != nil {
-			fmt.Fprintf(ioStreams.ErrOut, "%v\n", err)
-			return nil, cobra.ShellCompDirectiveNoFileComp
-		}
-
-		result, err := completer(f.Context(), manager)
-		if err != nil {
-			fmt.Fprintf(ioStreams.ErrOut, "%v\n", err)
-			return nil, cobra.ShellCompDirectiveNoFileComp
-		}
-
-		return util.FilterStringsByPrefix(toComplete, result), cobra.ShellCompDirectiveNoFileComp
-	}
-}
-
-func gardenFlagCompletionFunc(ctx context.Context, manager target.Manager) ([]string, error) {
-	return util.GardenNames(manager)
-}
-
-func projectFlagCompletionFunc(ctx context.Context, manager target.Manager) ([]string, error) {
-	return util.ProjectNamesForTarget(ctx, manager)
-}
-
-func seedFlagCompletionFunc(ctx context.Context, manager target.Manager) ([]string, error) {
-	return util.SeedNamesForTarget(ctx, manager)
-}
-
-func shootFlagCompletionFunc(ctx context.Context, manager target.Manager) ([]string, error) {
-	return util.ShootNamesForTarget(ctx, manager)
 }
 
 // addKlogFlags adds flags from k8s.io/klog
